Retire old SearchService class
